copy internal dvd player fixes from trunk to 0.20 fixes. Refs #2447, #2455, #2294
git-svn-id: http://svn.mythtv.org/svn/branches/release-0-20-fixes@11326 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/DVDRingBuffer.h
@@ -147,7 +147,7 @@ class DVDRingBufferPriv
     /// menu pkt pts is not reliable
     long long      menupktpts;
     int            curAudioTrack;
-    int            curSubtitleTrack;
+    int8_t         curSubtitleTrack;
     bool           autoselectaudio;
     bool           autoselectsubtitle;
     const char     *dvdname;
